Linux，增加dockerREST接口支持，并测试连接成功，daemon需要root权限，组件注册到mainW，网络接口更新，主要是REST的接口，下一步测试REST接口的收发

+++ core/helper/handle.cpp
@@ -6,7 +6,7 @@ void Handle::start(){
 	connect(this->socket, SIGNAL(disconnected()), this, SLOT(disconnected()));
 	connect(this->socket, SIGNAL(readyRead()), this, SLOT(read()));
 	qDebug() << "[DEBUG][HANDLE]" << QThread::currentThread() << "New connection";
-	emit this->information(QString("New connection: %1").arg((int)QThread::currentThreadId()));
+	emit this->information(QString("New connection: %1").arg((quint64)QThread::currentThreadId()));
 }
 
 void Handle::write(QJsonDocument doc){
@@ -39,7 +39,7 @@ void Handle::timerEvent(QTimerEvent *e){
 
 void Handle::disconnected(){
 	qDebug() << "[DEBUG][HANDLE]" << QThread::currentThread() << "Disconnected";
-	emit this->information(QString("Disconnect: %1").arg((int)QThread::currentThreadId()));
+	emit this->information(QString("Disconnect: %1").arg((quint64)QThread::currentThreadId()));
 	emit this->finished(QThread::currentThread());
 }
